% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter data for per line level analyses}
\usage{
filter_data(data_stride, data_animal_mean, data_animal_disp, threshold)
}
\arguments{
\item{data_stride}{a csv file containing data per stride}

\item{data_animal}{a csv file containing data per animal}
}
\value{
a list containing filtered data_per_stride and data_per_animal
}
\description{
Filters the strains by setting the minimum number of animals per
strain
}
\examples{
filter_data(data_stride = data_per_stride, data_animal = 
data_per_animal, threshold = 5)
}
