% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_circular_mean.R
\name{komp_circular_mean}
\alias{komp_circular_mean}
\title{Fit a circular-linear regression model}
\usage{
komp_circular_mean(data, CtrlStrain = "C57BL/6NJ", model)
}
\arguments{
\item{data}{data frame containing per animal data}

\item{CtrlStrain}{specify the control strain}

\item{model}{specify the model}
}
\value{
a list containing the results of the model
}
\description{
Fit a circular-linear regression model to circular posture
phenotypes
}
\examples{
komp_circular_mean(data, CtrlStrain = "C57BL/6NJ", model = "M3")
}
