% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_desc_plot.R
\name{komp_desc_plot}
\alias{komp_desc_plot}
\title{Linear gait phenotype distributions across strains}
\usage{
komp_desc_plot(df_animal, df_strain, phenotype, phenoname, align)
}
\arguments{
\item{df_animal}{data frame containing data at per animal level}

\item{df_strain}{data frame containing data at per strain level}

\item{phenotype}{phenotype label}

\item{align}{horizontal or vertical plot}
}
\value{
plot similar to Figure 2 in the manuscript
}
\description{
Creates a plot showing the phenotypic mean ± sd for all strains \code{\link{kompgait}}
}
\examples{
komp_desc_plot()
}
