% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_lda.R
\name{komp_lda}
\alias{komp_lda}
\title{Linear discriminant analysis}
\usage{
komp_lda(data_animal, Mutants, controlids.df)
}
\arguments{
\item{data_animal}{a csv file containing data per animal}

\item{Mutants}{a vector of mutants}

\item{controlids.df}{data frame containing appropriate controls}
}
\value{
a plot containing LDA space and gait feature
contributions to the LD axes and vignette plots
}
\description{
Project the data from the 9D gait space to a 2D space
}
\examples{
komp_lda(data_animal = data_animal, Mutants = 
c('Arpc5l-/+','Fam120b-/+'))
}
