% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_lmer_mean.R
\name{komp_lmer_mean}
\alias{komp_lmer_mean}
\title{Fit a LMM to linear gait traits}
\usage{
komp_lmer_mean(CtrlStrain, model)
}
\arguments{
\item{CtrlStrain}{specify the control strain}

\item{model}{specify the model}
}
\value{
a list containing the results of the model
}
\description{
Fit a linear mixed model to identify significant differences in trait's average value between
mutant and control strain
}
\examples{
komp_lmer_mean(CtrlStrain = "C57BL/6NJ", model = "M3")
}
