% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_outlier.R
\name{komp_outlier_strain}
\alias{komp_outlier_strain}
\title{Multivariate outlier detection}
\usage{
komp_outlier_strain(data)
}
\arguments{
\item{data}{data.frame containing data at the strain level}
}
\value{
a plot showing outlier mutant lines in red
}
\description{
Use a PCA-based projection outlier algorithm to identify outlier mutant lines
}
\examples{
komp_outlier_strain(data = data_per_strain)
}
