% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_select_controls.R
\name{komp_select_controls}
\alias{komp_select_controls}
\title{Select control animals from control strain}
\usage{
komp_select_controls(CtrlStrain, tw)
}
\arguments{
\item{CtrlStrain}{specify the control strain}

\item{tw}{specify the time period in days}
}
\value{
a data.frame containing containing Strain, MouseID, Type
}
\description{
Choose control animals from the control strain to be examined during a specific time frame (in
days)
}
\examples{
komp_select_controls(CtrlStrain="C57BL/6NJ", tw = 21)
}
