% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load and preprocess data}
\usage{
load_data(data, type)
}
\arguments{
\item{data}{a csv file containing data per stride}

\item{type}{linear or circular}
}
\value{
a list containing data_per_stride and data_per_animal
}
\description{
Loads the data and preprocess the features.
\enumerate{
\item Remove certain strains
\item Focus on certain speed bins
}
}
\examples{
load_data(data = data_per_stride)
}
