% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_plots.R
\name{plot_model_comparisons}
\alias{plot_model_comparisons}
\title{Make a bar plot showing the number of pink strains across different
analyses
A plot showing the number of pink strains found across different
analyses (M1-3, MvOut)}
\usage{
plot_model_comparisons(M1, M2, M3, MvOut)
}
\arguments{
\item{M1}{full trait name}

\item{M2}{trait name in the dataframe}

\item{M3}{effect size matrix - output of modeling function}

\item{MvOut}{lod scores matrix - output of modeling function}
}
\value{
a plot
}
\description{
Make a bar plot showing the number of pink strains across different
analyses
A plot showing the number of pink strains found across different
analyses (M1-3, MvOut)
}
\examples{
plot_model_comparisons(M1 = aveM1[[3]], M2 = aveM2[[3]], M3 = 
aveM3[[3]], MvOut = mv_outlier_strains[[2]])
}
