% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_plots.R
\name{plot_outlier_animal_comparison}
\alias{plot_outlier_animal_comparison}
\title{Make a vignette plot for the outlier animal}
\usage{
plot_outlier_animal_comparison(data_animal, df_out, Mutant, phenotypes)
}
\arguments{
\item{data_animal}{data_per_animal}

\item{df_out}{a dataframe containing outlier animals}

\item{Mutant}{Strain that the outlier animal belongs to}

\item{phenotypes}{body length adjusted phenotypes to use for
'    plotting}
}
\value{
a plot
}
\description{
A plot that compares the outlier animal to other animals in
belonging to the same strain tested within 3 weeks
analyses (M1-3, MvOut)
}
\examples{
plot_outlier_animal_comparison(data_animal = data_per_animal,
df_out = outlier_animals[[2]], Mutant = "Sfxn5-/+", phenotypes = 
c("Speed", "Limb Duty Factor", "Step Length")) 
}
