% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_comparison.R
\name{post_hoc_comparison}
\alias{post_hoc_comparison}
\title{Post hoc comparisons}
\usage{
post_hoc_comparison(M1, M2, M3, MvOut, type)
}
\arguments{
\item{M1}{a list obtained using komp_lmer_mean(data, CtrlStrain,
model = "M1")}

\item{M2}{a list obtained using komp_lmer_mean(data, CtrlStrain,
model = "M2")}

\item{M3}{a list obtained using komp_lmer_mean(data, CtrlStrain,
model = "M3")}

\item{MvOut}{a list obtained using komp_outlier_strain(data)}
}
\value{
Plot comparing the number of pink mutants
across different models and multivariate outliers
}
\description{
Comparing univariate and multivariate analyses to each other and
with existing findings
}
\examples{
post_hoc_comparison(M1, M2, M3, MvOut, type = "bar_plot")
}
