% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_plots.R
\name{UpSet_plot}
\alias{UpSet_plot}
\title{Make an UpSet plot}
\usage{
UpSet_plot(M1, M2, M3, MvOut)
}
\arguments{
\item{M1}{full trait name}

\item{M2}{trait name in the dataframe}

\item{M3}{effect size matrix - output of modeling function}

\item{MvOut}{lod scores matrix - output of modeling function}
}
\value{
a plot
}
\description{
A plot showing the common pink strains found across different
analyses (M1-3, MvOut)
}
\examples{
UpSet_plot(M1 = aveM1[[3]], M2 = aveM2[[3]], M3 = 
aveM3[[3]], MvOut = mv_outlier_strains[[2]])
}
