% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/komp_plots.R
\name{vignette_plot}
\alias{vignette_plot}
\title{Make a vignette plot of the results from per line analyses}
\usage{
vignette_plot(data_animal, phenotype, phenoname, esizeGen2, lodGen2, model)
}
\arguments{
\item{data_animal}{a csv file containing data per animal}

\item{phenotype}{full trait name}

\item{phenoname}{trait name in the dataframe}

\item{esizeGen2}{effect size matrix - output of modeling function}

\item{lodGen2}{lod scores matrix - output of modeling function}

\item{model}{character "M1-3"}
}
\value{
a plot
}
\description{
Compare mutant versus control for phenotypes identified as
significantly different by models M1-3
}
\examples{
vignette_plot(phenotype = "Stride Length", phenoname = "stride 
_length", esizeGen2 = dispM3[[9]], lodGen2 = dispM3[[10]], model = 
"M3")
}
